refactor(vm): the VM spec is rendered in one place
`build_vm_spec` and the three builders under it are what the manager hands
the driver, and a restore has to render the same devices a boot did — but
they lived at the bottom of the 1300-line boot task, which is also the file
a reader goes to for the READY gate and the resource handoff.

`sandbox/spec.rs` takes the four of them and their test; `checkpoint.rs`
and `pause.rs` name `spec::restore_spec` instead of reaching into `boot`.

Pure move — 221 insertions / 200 deletions, bodies byte-identical. Only
`build_vm_spec` changes at all, from private to `pub(super)` so the boot
task still calls it, and `boot.rs` drops the driver-type and
`KernelIpParam` imports that left with it.
